CQ-4344327 Updated AbstractBaseImpl to extend from AbstractComponentImpl
